requirements.txt
aiogram>=3.7.0
httpx>=0.27.0
aiosqlite>=0.20.0
python-dotenv>=1.0.1
openai>=1.30.0
aiohttp>=3.9.0
